test(datetime): migrate month-year-picker
